backend/pyproject.toml
[tool.poetry]
name = "ai-therapy-backend"
version = "0.1.0"
description = "AI Therapy Backend with FastAPI"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
openai = "^1.3.8"
gtts = "^2.4.0"
python-multipart = "^0.0.6"
pydantic = "^2.5.0"
google-cloud-texttospeech = "^2.18.0"
google-api-core = "^2.17.0"
google-generativeai = "^0.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 